+++ modules/core/src/core.module.ts
@@ -2,6 +2,7 @@ import { CommonModule } from "@angular/common";
 import { NgModule, ModuleWithProviders } from "@angular/core";
 import { ReactiveFormsModule } from "@angular/forms";
 import { DynamicFormService } from "./service/dynamic-form.service";
+import { DynamicFormValidationService } from "./service/dynamic-form-validation.service";
 import { DynamicIdDirective } from "./directive/dynamic-id.directive";
 import { DynamicTemplateDirective } from "./directive/dynamic-template.directive";
 
@@ -21,7 +22,8 @@ export class DynamicFormsCoreModule {
             ngModule: DynamicFormsCoreModule,
 
             providers: [
-                DynamicFormService
+                DynamicFormService,
+                DynamicFormValidationService
             ]
         };
     }
