+++ modules/core/src/service/dynamic-form-validation.service.spec.ts
@@ -1,0 +1,125 @@
+import { TestBed, inject } from "@angular/core/testing";
+import { ReactiveFormsModule, FormControl, NG_VALIDATORS, NG_ASYNC_VALIDATORS } from "@angular/forms";
+import { DynamicFormValidationService } from "./dynamic-form-validation.service";
+import { DynamicFormControlModel } from "../model/dynamic-form-control.model";
+import { DynamicInputModel } from "../model/input/dynamic-input.model";
+
+describe("DynamicFormValidationService test suite", () => {
+
+    let service: DynamicFormValidationService;
+
+    function testValidator() {
+        return {testValidator: {valid: true}};
+    }
+
+    function testAsyncValidator() {
+        return new Promise<boolean>(resolve => setTimeout(() => resolve(true), 0));
+    }
+
+    beforeEach(() => {
+
+        TestBed.configureTestingModule({
+            imports: [ReactiveFormsModule],
+            providers: [
+                DynamicFormValidationService,
+                {provide: NG_VALIDATORS, useValue: testValidator, multi: true},
+                {provide: NG_ASYNC_VALIDATORS, useValue: testAsyncValidator, multi: true}
+            ]
+        });
+    });
+
+    beforeEach(inject([DynamicFormValidationService],
+        (validationService: DynamicFormValidationService) => service = validationService));
+
+
+    it("should resolve validator by name correctly", () => {
+
+        expect(service.getValidator("required")).toBeDefined();
+        expect(service.getValidator("testValidator")).toBeDefined();
+    });
+
+
+    it("should resolve async validator by name correctly", () => {
+
+        expect(service.getAsyncValidator("testAsyncValidator")).toBeDefined();
+    });
+
+
+    it("should resolve validators from config correctly", () => {
+
+        let config: any = {required: null, maxLength: 7, minLength: 3},
+            validators = service.getValidators(config);
+
+        expect(validators.length === Object.keys(config).length).toBe(true);
+    });
+
+
+    it("should resolve custom validators from config correctly", () => {
+
+        let config: any = {required: null, maxLength: 7, testValidator: null},
+            validators = service.getValidators(config);
+
+        expect(validators.length === Object.keys(config).length).toBe(true);
+    });
+
+
+    it("should resolve custom validators from detailed config correctly", () => {
+
+        let config: any = {testValidator: {name: testValidator.name, args: null}},
+            validators = service.getValidators(config);
+
+        expect(validators.length === Object.keys(config).length).toBe(true);
+    });
+
+
+    it("should resolve custom async validators from config correctly", () => {
+
+        let config: any = {testAsyncValidator: null},
+            validators = service.getAsyncValidators(config);
+
+        expect(validators.length === Object.keys(config).length).toBe(true);
+    });
+
+
+    it("should throw when validator is not provided via NG_VALIDATORS", () => {
+
+        expect(() => service.getValidator("test", null))
+            .toThrow(new Error(`validator "test" is not provided via NG_VALIDATORS or NG_ASYNC_VALIDATORS`));
+    });
+
+
+    it("should create error messages correctly", () => {
+
+        let errorMessages,
+            testControl: FormControl = new FormControl(),
+            testModel: DynamicFormControlModel = new DynamicInputModel({
+                id: "testModel",
+                errorMessages: {
+                    required: "Field is required",
+                    custom1: "Field {{ id }} has a custom error",
+                    custom2: "Field has a custom error: {{ validator.param }}"
+                }
+            });
+
+        errorMessages = service.createErrorMessages(testControl, testModel);
+        expect(errorMessages.length).toBe(0);
+
+        testControl.setErrors({required: true});
+
+        errorMessages = service.createErrorMessages(testControl, testModel);
+        expect(errorMessages.length).toBe(1);
+        expect(errorMessages[0]).toEqual(testModel.errorMessages["required"]);
+
+        testControl.setErrors({custom1: true});
+
+        errorMessages = service.createErrorMessages(testControl, testModel);
+        expect(errorMessages.length).toBe(1);
+        expect(errorMessages[0]).toEqual(`Field ${testModel.id} has a custom error`);
+
+        testControl.setErrors({custom2: {param: 42}});
+
+        errorMessages = service.createErrorMessages(testControl, testModel);
+        expect(errorMessages.length).toBe(1);
+        expect(errorMessages[0]).toEqual("Field has a custom error: 42");
+    });
+});
