+++ modules/core/src/utils/validation.utils.spec.ts
@@ -5,34 +5,6 @@ import { DynamicInputModel } from "../model/input/dynamic-input.model";
 
 describe("Validation Utils test suite", () => {
 
-    it("should create error messages correctly", () => {
-
-        let errorMessages,
-            testControl: FormControl = new FormControl(),
-            testModel: DynamicFormControlModel = new DynamicInputModel({
-                id: "testModel",
-                errorMessages: {
-                    required: "Field is required",
-                    custom: "Field {{ id }} has a custom error"
-                }
-            });
-
-        errorMessages = ValidationUtils.createErrorMessages(testControl, testModel);
-        expect(errorMessages.length).toBe(0);
-
-        testControl.setErrors({required: true});
-
-        errorMessages = ValidationUtils.createErrorMessages(testControl, testModel);
-        expect(errorMessages.length).toBe(1);
-        expect(errorMessages[0]).toEqual(testModel.errorMessages["required"]);
-
-        testControl.setErrors({custom: true});
-
-        errorMessages = ValidationUtils.createErrorMessages(testControl, testModel);
-        expect(errorMessages.length).toBe(1);
-        expect(errorMessages[0]).toEqual(`Field ${testModel.id} has a custom error`);
-    });
-
     it("should detect a validator configuration correctly", () => {
 
         let testConfig1: any = {name: "test"},
