+++ modules/ui-basic/src/dynamic-form-basic.component.ts
@@ -10,6 +10,7 @@ import {
 } from "@angular/core";
 import { FormGroup } from "@angular/forms";
 import {
+    DynamicFormValidationService,
     DynamicFormControlModel,
     DynamicFormArrayGroupModel,
     DynamicFormControlComponent,
@@ -59,8 +60,8 @@ export class DynamicFormBasicComponent extends DynamicFormControlComponent imple
 
     type: BasicFormControlType | null;
 
-    constructor() {
-        super();
+    constructor(protected validationService: DynamicFormValidationService) {
+        super(validationService);
     }
 
     ngOnChanges(changes: SimpleChanges) {
